requirements.txt
Flask
flask-cors
flask-bcrypt
PyJWT
pymongo
python-dotenv
